The Normal distribution

The Normal distribution is ubiquitous in statistics, partially because of the central limit theorem, which states that sums of i.i.d. random variables eventually become Normal. Linear transformations of Normal random variables result in new random variables that are also Normal. If you are taking an intro stats course, you'll likely use the Normal distribution for Z-tests and in simple linear regression. Under regularity conditions, maximum likelihood estimators are asymptotically Normal. The Normal distribution is also called the gaussian distribution.

dist_normal(mu = 0, sigma = 1, mean = mu, sd = sigma)

Arguments

mu, mean

The mean (location parameter) of the distribution, which is also the mean of the distribution. Can be any real number.

sigma, sd

The standard deviation (scale parameter) of the distribution. Can be any positive number. If you would like a Normal distribution with variance \(\sigma^2\), be sure to take the square root, as this is a common source of errors.

Details

We recommend reading this documentation on https://pkg.mitchelloharawild.com/distributional/, where the math will render nicely.

In the following, let \(X\) be a Normal random variable with mean mu = \(\mu\) and standard deviation sigma = \(\sigma\).

Support: \(R\), the set of all real numbers

Mean: \(\mu\)

Variance: \(\sigma^2\)

Probability density function (p.d.f):

$$ f(x) = \frac{1}{\sqrt{2 \pi \sigma^2}} e^{-(x - \mu)^2 / 2 \sigma^2} $$

Cumulative distribution function (c.d.f):

The cumulative distribution function has the form

$$ F(t) = \int_{-\infty}^t \frac{1}{\sqrt{2 \pi \sigma^2}} e^{-(x - \mu)^2 / 2 \sigma^2} dx $$

but this integral does not have a closed form solution and must be approximated numerically. The c.d.f. of a standard Normal is sometimes called the "error function". The notation \(\Phi(t)\) also stands for the c.d.f. of a standard Normal evaluated at \(t\). Z-tables list the value of \(\Phi(t)\) for various \(t\).

Moment generating function (m.g.f):

$$ E(e^{tX}) = e^{\mu t + \sigma^2 t^2 / 2} $$

See also

Examples

dist <- dist_normal(mu = 1:5, sigma = 3)

dist
#> <distribution[5]>
#> [1] N(1, 9) N(2, 9) N(3, 9) N(4, 9) N(5, 9)
mean(dist)
#> [1] 1 2 3 4 5
variance(dist)
#> [1] 9 9 9 9 9
skewness(dist)
#> [1] 0 0 0 0 0
kurtosis(dist)
#> [1] 0 0 0 0 0

generate(dist, 10)
#> [[1]]
#>  [1] -0.9751405  0.9063073  4.0553655 -3.9563008  0.1469721  0.5949949
#>  [7] -3.8986873  1.0992223  0.2235669  0.7214432
#> 
#> [[2]]
#>  [1]  0.063909590 -3.478202585  3.865799391  4.634801299  1.227393560
#>  [6] -1.426557226  4.147522361  0.008738171  1.853152183  2.352227201
#> 
#> [[3]]
#>  [1]  7.901525  1.998106  2.808207  6.506206  2.253391  4.747666 -0.298780
#>  [8]  1.468851  2.883903  3.551551
#> 
#> [[4]]
#>  [1]  8.358468  8.330679  5.571204  2.527864  5.325788  4.229805 10.054959
#>  [8]  5.715203 11.080306  2.714303
#> 
#> [[5]]
#>  [1] 5.373908 1.968123 2.133772 2.915217 7.163452 4.757987 5.439322 9.325292
#>  [9] 4.197607 1.902055
#> 

density(dist, 2)
#> [1] 0.12579441 0.13298076 0.12579441 0.10648267 0.08065691
density(dist, 2, log = TRUE)
#> [1] -2.073106 -2.017551 -2.073106 -2.239773 -2.517551

cdf(dist, 4)
#> [1] 0.8413447 0.7475075 0.6305587 0.5000000 0.3694413

quantile(dist, 0.7)
#> [1] 2.573202 3.573202 4.573202 5.573202 6.573202