Create a distribution from p/d/q/r style functions

dist_wrap(dist, ..., package = "stats")

Arguments

dist
The name of the distribution used in the functions (name that is prefixed by p/d/q/r)

...
Named arguments used to parameterise the distribution.

package
The package from which the distribution is provided.

Details

If a distribution is not yet supported, you can vectorise p/d/q/r functions using this function. dist_wrap() stores the distributions parameters, and provides wrappers which call the appropriate p/d/q/r functions.

Using this function to wrap a distribution should only be done if the distribution is not yet available in this package. If you need a distribution which isn't in the package yet, consider making a request at https://github.com/mitchelloharawild/distributional/issues.

Examples

dist <- dist_wrap("norm", mean = 1:3, sd = c(3, 9, 2))

density(dist, 1) # dnorm()

#> [1] 0.13298076 0.04405414 0.12098536

cdf(dist, 4) # pnorm()

#> [1] 0.8413447 0.5879296 0.6914625

quantile(dist, 0.975) # qnorm()

#> [1]  6.879892 19.639676  6.919928

generate(dist, 10) # rnorm()

#> [[1]]
#>  [1] -0.2968551  4.4631311  5.5457843  1.6536806  2.0213429  4.8792756
#>  [7]  5.8200791 -0.4205119  2.0328075 -2.2873681
#> 
#> [[2]]
#>  [1] 11.3253522  4.6232238 -9.9705023 12.7965045  0.8269834 -1.8699909
#>  [7] -6.5214584 -6.7925362  4.3049609  6.5257643
#> 
#> [[3]]
#>  [1]  3.1476722  4.7217241  1.5509088 -0.6807630  3.4857620  0.1732843
#>  [7] -0.4618372  2.2537871  3.7132590  6.3780175
#> 

library(actuar)

#> 
#> Attaching package: ‘actuar’

#> The following object is masked from ‘package:grDevices’:
#> 
#>     cm

dist <- dist_wrap("invparalogis", package = "actuar", shape = 2, rate = 2)
density(dist, 1) # actuar::dinvparalogis()

#> [1] 0.512

cdf(dist, 4) # actuar::pinvparalogis()

#> [1] 0.9694675

quantile(dist, 0.975) # actuar::qinvparalogis()

#> [1] 4.429922

generate(dist, 10) # actuar::rinvparalogis()

#> [[1]]
#>  [1] 0.6608314 1.4285445 0.7360410 1.8259137 1.9844953 2.4594202 1.0218648
#>  [8] 1.1124026 3.8146147 0.3376756
#>