Construct hilo intervals

Class constructor function to help with manually creating hilo interval objects.

new_hilo(lower = double(), upper = double(), size = double())

Arguments

lower, upper

A numeric vector of values for lower and upper limits.

size

Size of the interval between [0, 100].

Value

A "hilo" vector

Author

Earo Wang & Mitchell O'Hara-Wild

Examples

new_hilo(lower = rnorm(10), upper = rnorm(10) + 5, size = 95)
#> <hilo[10]>
#>  [1] [ 0.2671142, 5.438906]95 [ 0.6247009, 4.726154]95 [-1.5705815, 5.616221]95
#>  [4] [ 1.2739797, 4.650031]95 [-0.6492718, 5.744182]95 [ 1.7295727, 3.962115]95
#>  [7] [-0.1105585, 4.019580]95 [-1.7302617, 3.562334]95 [ 1.0966358, 4.903387]95
#> [10] [ 1.5734263, 5.013213]95