The (non-central) Chi-Squared Distribution

Chi-square distributions show up often in frequentist settings as the sampling distribution of test statistics, especially in maximum likelihood estimation settings.

dist_chisq(df, ncp = 0)

Arguments

df

degrees of freedom (non-negative, but can be non-integer).

ncp

non-centrality parameter (non-negative).

Details

We recommend reading this documentation on https://pkg.mitchelloharawild.com/distributional/, where the math will render nicely.

In the following, let \(X\) be a \(\chi^2\) random variable with df = \(k\).

Support: \(R^+\), the set of positive real numbers

Mean: \(k\)

Variance: \(2k\)

Probability density function (p.d.f):

$$ f(x) = \frac{1}{\sqrt{2 \pi \sigma^2}} e^{-(x - \mu)^2 / 2 \sigma^2} $$

Cumulative distribution function (c.d.f):

The cumulative distribution function has the form

$$ F(t) = \int_{-\infty}^t \frac{1}{\sqrt{2 \pi \sigma^2}} e^{-(x - \mu)^2 / 2 \sigma^2} dx $$

but this integral does not have a closed form solution and must be approximated numerically. The c.d.f. of a standard normal is sometimes called the "error function". The notation \(\Phi(t)\) also stands for the c.d.f. of a standard normal evaluated at \(t\). Z-tables list the value of \(\Phi(t)\) for various \(t\).

Moment generating function (m.g.f):

$$ E(e^{tX}) = e^{\mu t + \sigma^2 t^2 / 2} $$

See also

Examples

dist <- dist_chisq(df = c(1,2,3,4,6,9))

dist
#> <distribution[6]>
#> [1] ᵪ²(1) ᵪ²(2) ᵪ²(3) ᵪ²(4) ᵪ²(6) ᵪ²(9)
mean(dist)
#> [1] 1 2 3 4 6 9
variance(dist)
#> [1]  2  4  6  8 12 18
skewness(dist)
#> [1] 2.828427 2.000000 1.632993 1.414214 1.154701 0.942809
kurtosis(dist)
#> [1] 12.000000  6.000000  4.000000  3.000000  2.000000  1.333333

generate(dist, 10)
#> [[1]]
#>  [1] 0.1066147 0.6872475 0.1028372 1.2121012 0.1021459 4.0981311 0.6977948
#>  [8] 0.9822044 1.0121547 0.5247478
#> 
#> [[2]]
#>  [1] 3.0911817 0.6409473 0.0833725 1.6684605 8.1632476 0.8862268 1.6191380
#>  [8] 4.1548161 3.6973979 1.6897608
#> 
#> [[3]]
#>  [1] 2.9863532 4.8700638 1.9178448 4.2287336 2.6018408 3.8655836 8.6996393
#>  [8] 2.5677337 1.3640800 0.2575386
#> 
#> [[4]]
#>  [1] 1.1817158 3.6536008 2.6537263 8.7801448 0.3974600 0.6803067 9.2643563
#>  [8] 4.3330761 2.0380241 1.6903988
#> 
#> [[5]]
#>  [1]  5.5952528 13.6344530  6.1016618  2.0604668  3.0749189  0.7702636
#>  [7]  3.4391337  5.4547475  4.3819753 18.2439398
#> 
#> [[6]]
#>  [1] 20.603831  9.217974  9.656846 10.224986 18.525539  1.896763  3.830344
#>  [8]  7.559584  2.449300  7.720162
#> 

density(dist, 2)
#> [1] 0.10377687 0.18393972 0.20755375 0.18393972 0.09196986 0.01581362
density(dist, 2, log = TRUE)
#> [1] -2.265512 -1.693147 -1.572365 -1.693147 -2.386294 -4.146884

cdf(dist, 4)
#> [1] 0.95449974 0.86466472 0.73853587 0.59399415 0.32332358 0.08858747

quantile(dist, 0.7)
#> [1]  1.074194  2.407946  3.664871  4.878433  7.231135 10.656372