The (non-central) Chi-Squared Distribution

dist_chisq(df, ncp = 0)

Arguments

df
degrees of freedom (non-negative, but can be non-integer).

ncp
non-centrality parameter (non-negative).

Details

Chi-square distributions show up often in frequentist settings as the sampling distribution of test statistics, especially in maximum likelihood estimation settings.

We recommend reading this documentation on https://pkg.mitchelloharawild.com/distributional/, where the math will render nicely.

In the following, let \(X\) be a \(\chi^2\) random variable with df = \(k\).

Support: \(R^+\), the set of positive real numbers

Mean: \(k\)

Variance: \(2k\)

Probability density function (p.d.f):

$$ f(x) = \frac{1}{\sqrt{2 \pi \sigma^2}} e^{-(x - \mu)^2 / 2 \sigma^2} $$

Cumulative distribution function (c.d.f):

The cumulative distribution function has the form

$$ F(t) = \int_{-\infty}^t \frac{1}{\sqrt{2 \pi \sigma^2}} e^{-(x - \mu)^2 / 2 \sigma^2} dx $$

but this integral does not have a closed form solution and must be approximated numerically. The c.d.f. of a standard normal is sometimes called the "error function". The notation \(\Phi(t)\) also stands for the c.d.f. of a standard normal evaluated at \(t\). Z-tables list the value of \(\Phi(t)\) for various \(t\).

Moment generating function (m.g.f):

$$ E(e^{tX}) = e^{\mu t + \sigma^2 t^2 / 2} $$

See also

Examples

dist <- dist_chisq(df = c(1,2,3,4,6,9))

dist

#> <distribution[6]>
#> [1] ᵪ²(1) ᵪ²(2) ᵪ²(3) ᵪ²(4) ᵪ²(6) ᵪ²(9)

mean(dist)

#> [1] 1 2 3 4 6 9

variance(dist)

#> [1]  2  4  6  8 12 18

skewness(dist)

#> [1] 2.828427 2.000000 1.632993 1.414214 1.154701 0.942809

kurtosis(dist)

#> [1] 12.000000  6.000000  4.000000  3.000000  2.000000  1.333333

generate(dist, 10)

#> [[1]]
#>  [1] 3.023731353 0.048417488 1.360303267 0.012077634 3.815762179 0.843698409
#>  [7] 0.045614425 0.233201302 0.002680881 0.077814108
#> 
#> [[2]]
#>  [1] 5.2140097 0.1138795 4.9075193 0.4674858 1.8660854 0.8597603 1.2525103
#>  [8] 1.3666645 1.3621329 0.6671215
#> 
#> [[3]]
#>  [1]  1.164510  4.038398 15.613134  2.493755  4.687919  4.938266  4.879185
#>  [8]  1.284802  6.480642  5.055150
#> 
#> [[4]]
#>  [1]  8.9435799  2.3703962  2.2069658  0.9125746  2.9868317 12.9969523
#>  [7]  0.8399264  2.8246700  0.6949930  3.0139185
#> 
#> [[5]]
#>  [1] 3.160282 6.399912 1.462012 7.077559 5.187542 1.753155 5.400800 2.341113
#>  [9] 5.728716 3.168700
#> 
#> [[6]]
#>  [1] 11.086178  3.053023  6.015485 11.267140  2.970658  5.795824  4.976395
#>  [8]  5.746530 17.014966 18.433337
#> 

density(dist, 2)

#> [1] 0.10377687 0.18393972 0.20755375 0.18393972 0.09196986 0.01581362

density(dist, 2, log = TRUE)

#> [1] -2.265512 -1.693147 -1.572365 -1.693147 -2.386294 -4.146884

cdf(dist, 4)

#> [1] 0.95449974 0.86466472 0.73853587 0.59399415 0.32332358 0.08858747

quantile(dist, 0.7)

#> [1]  1.074194  2.407946  3.664871  4.878433  7.231135 10.656372