The Gamma distribution

Several important distributions are special cases of the Gamma distribution. When the shape parameter is 1, the Gamma is an exponential distribution with parameter \(1/\beta\). When the \(shape = n/2\) and \(rate = 1/2\), the Gamma is a equivalent to a chi squared distribution with n degrees of freedom. Moreover, if we have \(X_1\) is \(Gamma(\alpha_1, \beta)\) and \(X_2\) is \(Gamma(\alpha_2, \beta)\), a function of these two variables of the form \(\frac{X_1}{X_1 + X_2}\) \(Beta(\alpha_1, \alpha_2)\). This last property frequently appears in another distributions, and it has extensively been used in multivariate methods. More about the Gamma distribution will be added soon.

dist_gamma(shape, rate, scale = 1/rate)

Arguments

shape, scale

shape and scale parameters. Must be positive, scale strictly.

rate

an alternative way to specify the scale.

Details

We recommend reading this documentation on https://pkg.mitchelloharawild.com/distributional/, where the math will render nicely.

In the following, let \(X\) be a Gamma random variable with parameters shape = \(\alpha\) and rate = \(\beta\).

Support: \(x \in (0, \infty)\)

Mean: \(\frac{\alpha}{\beta}\)

Variance: \(\frac{\alpha}{\beta^2}\)

Probability density function (p.m.f):

$$ f(x) = \frac{\beta^{\alpha}}{\Gamma(\alpha)} x^{\alpha - 1} e^{-\beta x} $$

Cumulative distribution function (c.d.f):

$$ f(x) = \frac{\Gamma(\alpha, \beta x)}{\Gamma{\alpha}} $$

Moment generating function (m.g.f):

$$ E(e^{tX}) = \Big(\frac{\beta}{ \beta - t}\Big)^{\alpha}, \thinspace t < \beta $$

See also

Examples

dist <- dist_gamma(shape = c(1,2,3,5,9,7.5,0.5), rate = c(0.5,0.5,0.5,1,2,1,1))

dist
#> <distribution[7]>
#> [1] Γ(1, 0.5) Γ(2, 0.5) Γ(3, 0.5) Γ(5, 1)   Γ(9, 2)   Γ(7.5, 1) Γ(0.5, 1)
mean(dist)
#> [1] 2.0 4.0 6.0 5.0 4.5 7.5 0.5
variance(dist)
#> [1]  4.00  8.00 12.00  5.00  2.25  7.50  0.50
skewness(dist)
#> [1] 2.0000000 1.4142136 1.1547005 0.8944272 0.6666667 0.7302967 2.8284271
kurtosis(dist)
#> [1]  6.0000000  3.0000000  2.0000000  1.2000000  0.6666667  0.8000000 12.0000000

generate(dist, 10)
#> [[1]]
#>  [1] 0.3925603 2.0570271 1.0696795 2.8332474 0.6706284 0.8148122 0.5296762
#>  [8] 1.7442028 3.6931241 4.7972481
#> 
#> [[2]]
#>  [1]  3.7593881  2.4144202  4.0305388  3.1933929  2.3175894  1.8869632
#>  [7]  2.6853001  1.9945000  0.5559899 14.8175811
#> 
#> [[3]]
#>  [1]  5.563762  5.204188  1.963848  4.753208 10.546455  7.779268  1.411438
#>  [8]  8.662279  2.010226  9.713115
#> 
#> [[4]]
#>  [1] 2.055536 4.688452 4.957677 9.120526 3.521305 3.550145 4.524105 1.412911
#>  [9] 7.968029 3.989366
#> 
#> [[5]]
#>  [1] 7.525165 2.794834 4.880426 4.052339 4.863990 3.041565 3.194554 4.880890
#>  [9] 9.236806 4.553431
#> 
#> [[6]]
#>  [1] 10.867471 13.643174  9.481535  9.620338  6.736293  7.549316  5.623518
#>  [8] 15.429489  4.830103 13.565856
#> 
#> [[7]]
#>  [1] 0.005756406 0.048147794 0.277324973 0.441455394 0.839214905 0.001741396
#>  [7] 0.010184018 0.577695746 0.899394875 0.061355303
#> 

density(dist, 2)
#> [1] 0.183939721 0.183939721 0.091969860 0.090223522 0.059540363 0.006545958
#> [7] 0.053990967
density(dist, 2, log = TRUE)
#> [1] -1.693147 -1.693147 -2.386294 -2.405465 -2.821101 -5.028908 -2.918939

cdf(dist, 4)
#> [1] 0.8646647 0.5939942 0.3233236 0.3711631 0.4074527 0.0762173 0.9953223

quantile(dist, 0.7)
#> [1] 2.4079456 4.8784330 7.2311353 5.8903613 5.1503385 8.6608472 0.5370971