The Gamma distribution

Several important distributions are special cases of the Gamma distribution. When the shape parameter is 1, the Gamma is an exponential distribution with parameter \(1/\beta\). When the \(shape = n/2\) and \(rate = 1/2\), the Gamma is a equivalent to a chi squared distribution with n degrees of freedom. Moreover, if we have \(X_1\) is \(Gamma(\alpha_1, \beta)\) and \(X_2\) is \(Gamma(\alpha_2, \beta)\), a function of these two variables of the form \(\frac{X_1}{X_1 + X_2}\) \(Beta(\alpha_1, \alpha_2)\). This last property frequently appears in another distributions, and it has extensively been used in multivariate methods. More about the Gamma distribution will be added soon.

dist_gamma(shape, rate = 1/scale, scale = 1/rate)

Arguments

shape, scale

shape and scale parameters. Must be positive, scale strictly.

rate

an alternative way to specify the scale.

Details

We recommend reading this documentation on pkgdown which renders math nicely. https://pkg.mitchelloharawild.com/distributional/reference/dist_gamma.html

In the following, let \(X\) be a Gamma random variable with parameters shape = \(\alpha\) and rate = \(\beta\).

Support: \(x \in (0, \infty)\)

Mean: \(\frac{\alpha}{\beta}\)

Variance: \(\frac{\alpha}{\beta^2}\)

Probability density function (p.m.f):

$$ f(x) = \frac{\beta^{\alpha}}{\Gamma(\alpha)} x^{\alpha - 1} e^{-\beta x} $$

Cumulative distribution function (c.d.f):

$$ f(x) = \frac{\Gamma(\alpha, \beta x)}{\Gamma{\alpha}} $$

Moment generating function (m.g.f):

$$ E(e^{tX}) = \Big(\frac{\beta}{ \beta - t}\Big)^{\alpha}, \thinspace t < \beta $$

See also

Examples

dist <- dist_gamma(shape = c(1,2,3,5,9,7.5,0.5), rate = c(0.5,0.5,0.5,1,2,1,1))

dist
#> <distribution[7]>
#> [1] Γ(1, 0.5) Γ(2, 0.5) Γ(3, 0.5) Γ(5, 1)   Γ(9, 2)   Γ(7.5, 1) Γ(0.5, 1)
mean(dist)
#> [1] 2.0 4.0 6.0 5.0 4.5 7.5 0.5
variance(dist)
#> [1]  4.00  8.00 12.00  5.00  2.25  7.50  0.50
skewness(dist)
#> [1] 2.0000000 1.4142136 1.1547005 0.8944272 0.6666667 0.7302967 2.8284271
kurtosis(dist)
#> [1]  6.0000000  3.0000000  2.0000000  1.2000000  0.6666667  0.8000000 12.0000000

generate(dist, 10)
#> [[1]]
#>  [1] 0.94448290 1.16788851 0.19099072 3.27286527 0.01536563 3.00866218
#>  [7] 0.46505369 0.73911704 0.46497507 1.11581299
#> 
#> [[2]]
#>  [1] 0.7088386 3.9704497 7.3330652 1.2519780 4.4991009 4.4975226 1.6019558
#>  [8] 4.5599795 1.8140502 1.3896288
#> 
#> [[3]]
#>  [1] 6.316528 8.275151 9.521868 9.428109 2.194702 7.129521 3.237007 9.837764
#>  [9] 4.225715 4.738433
#> 
#> [[4]]
#>  [1] 2.526915 4.052192 4.999339 5.915749 4.746961 8.248155 3.364596 2.705230
#>  [9] 7.447256 3.611930
#> 
#> [[5]]
#>  [1] 3.156328 2.536666 6.980342 3.363181 7.038153 4.264674 6.575509 4.649641
#>  [9] 4.859170 3.107934
#> 
#> [[6]]
#>  [1] 5.269189 5.786172 4.954809 4.815212 5.422151 9.771682 6.630607 4.834448
#>  [9] 6.320563 5.675854
#> 
#> [[7]]
#>  [1] 1.9382686504 0.1724916789 0.1175260944 1.6516119584 0.4508443426
#>  [6] 0.8087757069 0.7064180776 0.0006658913 0.0049007131 0.2980294873
#> 

density(dist, 2)
#> [1] 0.183939721 0.183939721 0.091969860 0.090223522 0.059540363 0.006545958
#> [7] 0.053990967
density(dist, 2, log = TRUE)
#> [1] -1.693147 -1.693147 -2.386294 -2.405465 -2.821101 -5.028908 -2.918939

cdf(dist, 4)
#> [1] 0.8646647 0.5939942 0.3233236 0.3711631 0.4074527 0.0762173 0.9953223

quantile(dist, 0.7)
#> [1] 2.4079456 4.8784330 7.2311353 5.8903613 5.1503385 8.6608472 0.5370971