The Gamma distribution

Several important distributions are special cases of the Gamma distribution. When the shape parameter is 1, the Gamma is an exponential distribution with parameter \(1/\beta\). When the \(shape = n/2\) and \(rate = 1/2\), the Gamma is a equivalent to a chi squared distribution with n degrees of freedom. Moreover, if we have \(X_1\) is \(Gamma(\alpha_1, \beta)\) and \(X_2\) is \(Gamma(\alpha_2, \beta)\), a function of these two variables of the form \(\frac{X_1}{X_1 + X_2}\) \(Beta(\alpha_1, \alpha_2)\). This last property frequently appears in another distributions, and it has extensively been used in multivariate methods. More about the Gamma distribution will be added soon.

dist_gamma(shape, rate, scale = 1/rate)

Arguments

shape, scale

shape and scale parameters. Must be positive, scale strictly.

rate

an alternative way to specify the scale.

Details

We recommend reading this documentation on https://pkg.mitchelloharawild.com/distributional/, where the math will render nicely.

In the following, let \(X\) be a Gamma random variable with parameters shape = \(\alpha\) and rate = \(\beta\).

Support: \(x \in (0, \infty)\)

Mean: \(\frac{\alpha}{\beta}\)

Variance: \(\frac{\alpha}{\beta^2}\)

Probability density function (p.m.f):

$$ f(x) = \frac{\beta^{\alpha}}{\Gamma(\alpha)} x^{\alpha - 1} e^{-\beta x} $$

Cumulative distribution function (c.d.f):

$$ f(x) = \frac{\Gamma(\alpha, \beta x)}{\Gamma{\alpha}} $$

Moment generating function (m.g.f):

$$ E(e^{tX}) = \Big(\frac{\beta}{ \beta - t}\Big)^{\alpha}, \thinspace t < \beta $$

See also

Examples

dist <- dist_gamma(shape = c(1,2,3,5,9,7.5,0.5), rate = c(0.5,0.5,0.5,1,2,1,1))

dist
#> <distribution[7]>
#> [1] Γ(1, 0.5) Γ(2, 0.5) Γ(3, 0.5) Γ(5, 1)   Γ(9, 2)   Γ(7.5, 1) Γ(0.5, 1)
mean(dist)
#> [1] 2.0 4.0 6.0 5.0 4.5 7.5 0.5
variance(dist)
#> [1]  4.00  8.00 12.00  5.00  2.25  7.50  0.50
skewness(dist)
#> [1] 2.0000000 1.4142136 1.1547005 0.8944272 0.6666667 0.7302967 2.8284271
kurtosis(dist)
#> [1]  6.0000000  3.0000000  2.0000000  1.2000000  0.6666667  0.8000000 12.0000000

generate(dist, 10)
#> [[1]]
#>  [1] 1.09151677 0.56131180 0.09353204 0.40435461 3.40747023 2.15390359
#>  [7] 6.51187518 7.41812929 3.56274747 1.33050974
#> 
#> [[2]]
#>  [1] 10.6266955  3.2277882  0.7163586  0.3818369  3.1841318 13.6396875
#>  [7]  2.8099091  3.6857501  5.5754641  1.1196156
#> 
#> [[3]]
#>  [1]  0.6280509 10.3712049 18.8556150  5.7181777  3.0552349  4.9124656
#>  [7]  6.0932573  6.1025130 10.3560330  2.8340610
#> 
#> [[4]]
#>  [1] 3.455234 7.514794 2.630157 3.904155 5.449911 8.496484 7.864584 3.785039
#>  [9] 5.255440 3.672237
#> 
#> [[5]]
#>  [1] 6.253478 2.581500 3.207688 1.689477 4.155992 5.055745 2.891685 3.933688
#>  [9] 4.462094 6.330628
#> 
#> [[6]]
#>  [1]  8.940611  8.131849  5.144869 10.355557  4.477071  8.163972  8.793458
#>  [8]  7.730500  6.581083  6.958249
#> 
#> [[7]]
#>  [1] 0.37638058 0.01552343 0.42259779 0.44481944 0.23820153 0.11660630
#>  [7] 0.08644417 0.40702216 0.15272538 0.61136044
#> 

density(dist, 2)
#> [1] 0.183939721 0.183939721 0.091969860 0.090223522 0.059540363 0.006545958
#> [7] 0.053990967
density(dist, 2, log = TRUE)
#> [1] -1.693147 -1.693147 -2.386294 -2.405465 -2.821101 -5.028908 -2.918939

cdf(dist, 4)
#> [1] 0.8646647 0.5939942 0.3233236 0.3711631 0.4074527 0.0762173 0.9953223

quantile(dist, 0.7)
#> [1] 2.4079456 4.8784330 7.2311353 5.8903613 5.1503385 8.6608472 0.5370971