The Gamma distribution

dist_gamma(shape, rate)

Arguments

shape

shape and scale parameters. Must be positive, scale strictly.

rate

an alternative way to specify the scale.

Details

Several important distributions are special cases of the Gamma distribution. When the shape parameter is 1, the Gamma is an exponential distribution with parameter \(1/\beta\). When the \(shape = n/2\) and \(rate = 1/2\), the Gamma is a equivalent to a chi squared distribution with n degrees of freedom. Moreover, if we have \(X_1\) is \(Gamma(\alpha_1, \beta)\) and \(X_2\) is \(Gamma(\alpha_2, \beta)\), a function of these two variables of the form \(\frac{X_1}{X_1 + X_2}\) \(Beta(\alpha_1, \alpha_2)\). This last property frequently appears in another distributions, and it has extensively been used in multivariate methods. More about the Gamma distribution will be added soon.

We recommend reading this documentation on https://pkg.mitchelloharawild.com/distributional/, where the math will render nicely.

In the following, let \(X\) be a Gamma random variable with parameters shape = \(\alpha\) and rate = \(\beta\).

Support: \(x \in (0, \infty)\)

Mean: \(\frac{\alpha}{\beta}\)

Variance: \(\frac{\alpha}{\beta^2}\)

Probability density function (p.m.f):

$$ f(x) = \frac{\beta^{\alpha}}{\Gamma(\alpha)} x^{\alpha - 1} e^{-\beta x} $$

Cumulative distribution function (c.d.f):

$$ f(x) = \frac{\Gamma(\alpha, \beta x)}{\Gamma{\alpha}} $$

Moment generating function (m.g.f):

$$ E(e^{tX}) = \Big(\frac{\beta}{ \beta - t}\Big)^{\alpha}, \thinspace t < \beta $$

See also

Examples

dist <- dist_gamma(shape = c(1,2,3,5,9,7.5,0.5), rate = c(0.5,0.5,0.5,1,2,1,1))

dist
#> <distribution[7]>
#> [1] Γ(1, 0.5) Γ(2, 0.5) Γ(3, 0.5) Γ(5, 1)   Γ(9, 2)   Γ(7.5, 1) Γ(0.5, 1)
mean(dist)
#> [1] 2.0 4.0 6.0 5.0 4.5 7.5 0.5
variance(dist)
#> [1]  4.00  8.00 12.00  5.00  2.25  7.50  0.50
skewness(dist)
#> [1] 2.0000000 1.4142136 1.1547005 0.8944272 0.6666667 0.7302967 2.8284271
kurtosis(dist)
#> [1]  6.0000000  3.0000000  2.0000000  1.2000000  0.6666667  0.8000000 12.0000000

generate(dist, 10)
#> [[1]]
#>  [1] 0.18318971 4.53335534 0.05740829 3.76293238 1.13729098 2.89642649
#>  [7] 0.09502289 1.07370025 2.81202889 1.57017983
#> 
#> [[2]]
#>  [1] 3.320552 7.782896 1.894936 6.362897 2.711260 1.848043 6.990566 2.421463
#>  [9] 1.407190 3.052881
#> 
#> [[3]]
#>  [1] 10.931732 15.956400  7.204284  7.026131  5.558358  3.553256  3.575093
#>  [8]  3.692849  8.542365  1.418457
#> 
#> [[4]]
#>  [1] 2.593961 5.761683 3.741532 1.770919 4.485481 4.273706 6.205996 2.975264
#>  [9] 7.361879 6.426164
#> 
#> [[5]]
#>  [1] 4.299684 3.346362 4.313597 3.695464 2.475638 2.392844 5.374113 8.525388
#>  [9] 6.965458 2.738229
#> 
#> [[6]]
#>  [1]  6.927597  4.893213 10.852856  5.014793  5.748474 10.581604  2.528426
#>  [8]  7.249896 13.089539  6.077930
#> 
#> [[7]]
#>  [1] 0.4789826 0.9066987 0.5163134 0.2067469 0.2825865 0.6318903 0.1744980
#>  [8] 0.1016402 0.8068738 0.9459037
#> 

density(dist, 2)
#> [1] 0.183939721 0.183939721 0.091969860 0.090223522 0.059540363 0.006545958
#> [7] 0.053990967
density(dist, 2, log = TRUE)
#> [1] -1.693147 -1.693147 -2.386294 -2.405465 -2.821101 -5.028908 -2.918939

cdf(dist, 4)
#> [1] 0.8646647 0.5939942 0.3233236 0.3711631 0.4074527 0.0762173 0.9953223

quantile(dist, 0.7)
#> [1] 2.4079456 4.8784330 7.2311353 5.8903613 5.1503385 8.6608472 0.5370971