The Gumbel distribution

dist_gumbel(alpha, scale)

Arguments

alpha
location parameter.

scale
parameter. Must be strictly positive.

Details

The Gumbel distribution is a special case of the Generalized Extreme Value distribution, obtained when the GEV shape parameter \(\xi\) is equal to 0. It may be referred to as a type I extreme value distribution.

We recommend reading this documentation on https://pkg.mitchelloharawild.com/distributional/, where the math will render nicely.

In the following, let \(X\) be a Gumbel random variable with location parameter mu = \(\mu\), scale parameter sigma = \(\sigma\).

Support: \(R\), the set of all real numbers.

Mean: \(\mu + \sigma\gamma\), where \(\gamma\) is Euler's constant, approximately equal to 0.57722.

Median: \(\mu - \sigma\ln(\ln 2)\).

Variance: \(\sigma^2 \pi^2 / 6\).

Probability density function (p.d.f):

$$f(x) = \sigma ^ {-1} \exp[-(x - \mu) / \sigma]% \exp\{-\exp[-(x - \mu) / \sigma] \}$$ for \(x\) in \(R\), the set of all real numbers.

Cumulative distribution function (c.d.f):

In the \(\xi = 0\) (Gumbel) special case $$F(x) = \exp\{-\exp[-(x - \mu) / \sigma] \}$$ for \(x\) in \(R\), the set of all real numbers.

See also

Examples

dist <- dist_gumbel(alpha = c(0.5, 1, 1.5, 3), scale = c(2, 2, 3, 4))

dist

#> <distribution[4]>
#> [1] Gumbel(0.5, 2) Gumbel(1, 2)   Gumbel(1.5, 3) Gumbel(3, 4)  

mean(dist)

#> [1] 1.654431 2.154431 3.231647 5.308863

variance(dist)

#> [1]  6.579736  6.579736 14.804407 26.318945

skewness(dist)

#> [1] 1.139547 1.139547 1.139547 1.139547

kurtosis(dist)

#> [1] 2.4 2.4 2.4 2.4

generate(dist, 10)

#> [[1]]
#>  [1] 0.00987875 0.56468937 1.28090665 0.73315464 3.42096996 1.22048248
#>  [7] 0.32607253 4.54169956 6.19758786 3.08902955
#> 
#> [[2]]
#>  [1] 0.8160622 3.2259478 2.0300921 3.3499724 1.3217829 1.1965709 2.5290666
#>  [8] 0.6719784 5.0159113 1.2291148
#> 
#> [[3]]
#>  [1]  3.46640841 -0.05981758  4.10383401  2.01245359  3.31474676  7.20891982
#>  [7] -2.55251814  3.31511793 -1.58000921  6.53569374
#> 
#> [[4]]
#>  [1]  6.9409190  6.0148134  4.4622217 10.5195918  1.5017798 -0.9385159
#>  [7]  0.7447138  8.8035333  7.0147365  5.1709983
#> 

density(dist, 2)

#> [1] 0.14726616 0.16535215 0.12102469 0.08889319

density(dist, 2, log = TRUE)

#> [1] -1.915514 -1.799678 -2.111761 -2.420320

cdf(dist, 4)

#> [1] 0.8404869 0.8000107 0.6475248 0.4589561

quantile(dist, 0.7)

#> [1] 2.561861 3.061861 4.592791 7.123722