The Inverse Exponential distribution

The Inverse Exponential distribution is used to model the reciprocal of exponentially distributed variables.

dist_inverse_exponential(rate)

Arguments

rate

an alternative way to specify the scale.

Details

We recommend reading this documentation on pkgdown which renders math nicely. https://pkg.mitchelloharawild.com/distributional/reference/dist_inverse_exponential.html

In the following, let \(X\) be an Inverse Exponential random variable with parameter rate = \(\lambda\).

Support: \(x > 0\)

Mean: Does not exist, returns NA

Variance: Does not exist, returns NA

Probability density function (p.d.f):

$$ f(x) = \frac{\lambda}{x^2} e^{-\lambda/x} $$

Cumulative distribution function (c.d.f):

$$ F(x) = e^{-\lambda/x} $$

Quantile function (inverse c.d.f):

$$ F^{-1}(p) = -\frac{\lambda}{\log(p)} $$

Moment generating function (m.g.f):

Does not exist (divergent integral).

See also

Examples

dist <- dist_inverse_exponential(rate = 1:5)
dist
#> <distribution[5]>
#> [1] InvExp(1) InvExp(2) InvExp(3) InvExp(4) InvExp(5)

mean(dist)
#> [1] NA NA NA NA NA
variance(dist)
#> [1] NA NA NA NA NA
support(dist)
#> <support_region[5]>
#> [1] (0,Inf) (0,Inf) (0,Inf) (0,Inf) (0,Inf)
generate(dist, 10)
#> [[1]]
#>  [1]  0.3375619  2.5245630  0.6309282 11.7820011  1.6847930  2.2766113
#>  [7]  2.1911598  0.3027820  0.8040568  0.4857418
#> 
#> [[2]]
#>  [1] 0.2886925 4.7558354 0.3954956 0.5715698 0.3434380 0.4930730 0.2594594
#>  [8] 0.6504981 9.7058283 1.3490965
#> 
#> [[3]]
#>  [1] 1.6531835 0.8085102 0.4756717 1.3677649 1.4778732 5.9705618 0.3649916
#>  [8] 0.7051300 1.3205321 0.7496672
#> 
#> [[4]]
#>  [1] 0.4230830 0.6494806 1.1576601 2.2664585 1.0650897 0.4635062 0.2782364
#>  [8] 0.9642015 0.4132235 2.9595911
#> 
#> [[5]]
#>  [1]  0.76337295  0.10500863  0.09930901  2.20309431  0.06283230  8.06907211
#>  [7]  0.22794054  0.71541982  0.40672713 17.78734250
#> 

density(dist, 2)
#> [1] 0.15163266 0.09735010 0.07054014 0.05515606 0.04524187
density(dist, 2, log = TRUE)
#> [1] -1.886294 -2.329442 -2.651573 -2.897589 -3.095732

cdf(dist, 4)
#> [1] 0.7788008 0.8824969 0.9200444 0.9394131 0.9512294

quantile(dist, 0.7)
#> [1] 2.8036733 1.4018366 0.9345578 0.7009183 0.5607347