The Inverse Gaussian distribution

dist_inverse_gaussian(mean, shape)

Arguments

mean, shape

parameters. Must be strictly positive. Infinite values are supported.

Details

The inverse Gaussian distribution (also known as the Wald distribution) is commonly used to model positive-valued data, particularly in contexts involving first passage times and reliability analysis.

We recommend reading this documentation on pkgdown which renders math nicely. https://pkg.mitchelloharawild.com/distributional/reference/dist_inverse_gaussian.html

In the following, let \(X\) be an Inverse Gaussian random variable with parameters mean = \(\mu\) and shape = \(\lambda\).

Support: \((0, \infty)\)

Mean: \(\mu\)

Variance: \(\frac{\mu^3}{\lambda}\)

Probability density function (p.d.f):

$$ f(x) = \sqrt{\frac{\lambda}{2\pi x^3}} \exp\left(-\frac{\lambda(x - \mu)^2}{2\mu^2 x}\right) $$

Cumulative distribution function (c.d.f):

$$ F(x) = \Phi\left(\sqrt{\frac{\lambda}{x}} \left(\frac{x}{\mu} - 1\right)\right) + \exp\left(\frac{2\lambda}{\mu}\right) \Phi\left(-\sqrt{\frac{\lambda}{x}} \left(\frac{x}{\mu} + 1\right)\right) $$

where \(\Phi\) is the standard normal c.d.f.

Moment generating function (m.g.f):

$$ E(e^{tX}) = \exp\left(\frac{\lambda}{\mu} \left(1 - \sqrt{1 - \frac{2\mu^2 t}{\lambda}}\right)\right) $$

for \(t < \frac{\lambda}{2\mu^2}\).

Skewness: \(3\sqrt{\frac{\mu}{\lambda}}\)

Excess Kurtosis: \(\frac{15\mu}{\lambda}\)

Quantiles: No closed-form expression, approximated numerically.

See also

Examples

dist <- dist_inverse_gaussian(mean = c(1,1,1,3,3), shape = c(0.2, 1, 3, 0.2, 1))
dist
#> <distribution[5]>
#> [1] IG(1, 0.2) IG(1, 1)   IG(1, 3)   IG(3, 0.2) IG(3, 1)  

mean(dist)
#> [1] 1 1 1 3 3
variance(dist)
#> [1]   5.0000000   1.0000000   0.3333333 135.0000000  27.0000000
support(dist)
#> <support_region[5]>
#> [1] (0,Inf) (0,Inf) (0,Inf) (0,Inf) (0,Inf)
generate(dist, 10)
#> [[1]]
#>  [1] 0.13249064 0.17335821 0.07694211 0.06082518 0.26330531 0.51860264
#>  [7] 6.27727743 3.80183264 0.18108779 0.29646195
#> 
#> [[2]]
#>  [1] 0.7152349 0.4591621 0.2984863 0.1783244 0.9014937 0.2529007 0.7395735
#>  [8] 0.8480113 1.9336480 0.1582204
#> 
#> [[3]]
#>  [1] 0.8998174 0.7689833 0.3547810 0.5002499 0.8231370 0.2442181 1.6859874
#>  [8] 0.6698564 0.3978965 0.7663033
#> 
#> [[4]]
#>  [1] 0.1695905 0.4093417 1.2103417 7.5988385 0.3068940 1.1667255 0.5866491
#>  [8] 0.2633586 1.3794918 0.6761230
#> 
#> [[5]]
#>  [1] 4.8027588 0.4237191 2.7058244 5.3315717 1.2369543 1.0060668 0.9426155
#>  [8] 8.4017231 1.4710383 0.9426707
#> 

density(dist, 2)
#> [1] 0.06000195 0.10984782 0.11539974 0.06272885 0.13718333
density(dist, 2, log = TRUE)
#> [1] -2.813378 -2.208659 -2.159353 -2.768934 -1.986437

cdf(dist, 4)
#> [1] 0.9454196 0.9790764 0.9983186 0.8735512 0.8031695

quantile(dist, 0.7)
#> [1] 0.6758386 1.0851197 1.1505484 1.0143030 2.5216357