The Laplace distribution

The Laplace distribution, also known as the double exponential distribution, is a continuous probability distribution that is symmetric around its location parameter.

dist_laplace(mu, sigma)

Arguments

mu

The location parameter (mean) of the Laplace distribution.

sigma

The positive scale parameter of the Laplace distribution.

Details

We recommend reading this documentation on https://pkg.mitchelloharawild.com/distributional/, where the math will render nicely.

In the following, let \(X\) be a Laplace random variable with location parameter mu = \(\mu\) and scale parameter sigma = \(\sigma\).

Support: \(R\), the set of all real numbers

Mean: \(\mu\)

Variance: \(2\sigma^2\)

Probability density function (p.d.f):

$$ f(x) = \frac{1}{2\sigma} \exp\left(-\frac{|x - \mu|}{\sigma}\right) $$

Cumulative distribution function (c.d.f):

$$ F(x) = \begin{cases} \frac{1}{2} \exp\left(\frac{x - \mu}{\sigma}\right) & \text{if } x < \mu \\ 1 - \frac{1}{2} \exp\left(-\frac{x - \mu}{\sigma}\right) & \text{if } x \geq \mu \end{cases} $$

Moment generating function (m.g.f):

$$ E(e^{tX}) = \frac{\exp(\mu t)}{1 - \sigma^2 t^2} \text{ for } |t| < \frac{1}{\sigma} $$

See also

Examples

dist <- dist_laplace(mu = c(0, 2, -1), sigma = c(1, 2, 0.5))

dist
#> <distribution[3]>
#> [1] Laplace(0, 1)    Laplace(2, 2)    Laplace(-1, 0.5)
mean(dist)
#> [1]  0  2 -1
variance(dist)
#> [1] 2.0 8.0 0.5
skewness(dist)
#> [1] 0 0 0
kurtosis(dist)
#> [1] 3 3 3

generate(dist, 10)
#> [[1]]
#>  [1] -0.02224686 -0.63015012  3.64099491  0.38280620  0.13190520  0.23312344
#>  [7] -0.95507441 -0.29660294 -1.42628736 -0.23392835
#> 
#> [[2]]
#>  [1] -3.0902683  2.5990092  2.0953143  1.5189393  4.8194257  0.8984396
#>  [7]  2.6140190  1.0633507 -0.4644326  4.0488765
#> 
#> [[3]]
#>  [1] -0.6621109 -0.7801682 -1.4454633  1.0659041 -0.7938208 -0.6456017
#>  [7] -0.8284758 -1.7795333 -2.1532482 -0.9464356
#> 

density(dist, 0)
#> [1] 0.50000000 0.09196986 0.13533528
density(dist, 0, log = TRUE)
#> [1] -0.6931472 -2.3862944 -2.0000000

cdf(dist, 1)
#> [1] 0.8160603 0.3032653 0.9908422

quantile(dist, 0.7)
#> [1]  0.5108256  3.0216512 -0.7445872