The Negative Binomial distribution

dist_negative_binomial(size, prob)

Arguments

size
target for number of successful trials, or dispersion parameter (the shape parameter of the gamma mixing distribution). Must be strictly positive, need not be integer.

prob
probability of success in each trial. 0 < prob <= 1.

Details

A generalization of the geometric distribution. It is the number of failures in a sequence of i.i.d. Bernoulli trials before a specified number of successes (size) occur. The probability of success in each trial is given by prob.

We recommend reading this documentation on https://pkg.mitchelloharawild.com/distributional/, where the math will render nicely.

In the following, let \(X\) be a Negative Binomial random variable with success probability prob = \(p\) and the number of successes size = \(r\).

Support: \(\{0, 1, 2, 3, ...\}\)

Mean: \(\frac{p r}{1-p}\)

Variance: \(\frac{pr}{(1-p)^2}\)

Probability mass function (p.m.f):

$$ f(k) = {k + r - 1 \choose k} \cdot (1-p)^r p^k $$

Cumulative distribution function (c.d.f):

Too nasty, omitted.

Moment generating function (m.g.f):

$$ \left(\frac{1-p}{1-pe^t}\right)^r, t < -\log p $$

See also

Examples

dist <- dist_negative_binomial(size = 10, prob = 0.5)

dist

#> <distribution[1]>
#> [1] NB(10, 0.5)

mean(dist)

#> [1] 10

variance(dist)

#> [1] 20

skewness(dist)

#> [1] 0.6708204

kurtosis(dist)

#> [1] 0.6125

generate(dist, 10)

#> [[1]]
#>  [1]  8  9 10 11 10 14 14  5  4  7
#> 

density(dist, 2)

#> [1] 0.01342773

density(dist, 2, log = TRUE)

#> [1] -4.310433

cdf(dist, 4)

#> [1] 0.08978271

quantile(dist, 0.7)

#> [1] 12