The Pareto Distribution

The Pareto distribution is a power-law probability distribution commonly used in actuarial science to model loss severity and in economics to model income distributions and firm sizes.

dist_pareto(shape, scale)

Arguments

shape, scale

parameters. Must be strictly positive.

Details

We recommend reading this documentation on pkgdown which renders math nicely. https://pkg.mitchelloharawild.com/distributional/reference/dist_pareto.html

In the following, let \(X\) be a Pareto random variable with parameters shape = \(\alpha\) and scale = \(\theta\).

Support: \((0, \infty)\)

Mean: \(\frac{\theta}{\alpha - 1}\) for \(\alpha > 1\), undefined otherwise

Variance: \(\frac{\alpha\theta^2}{(\alpha - 1)^2(\alpha - 2)}\) for \(\alpha > 2\), undefined otherwise

Probability density function (p.d.f):

$$ f(x) = \frac{\alpha\theta^\alpha}{(x + \theta)^{\alpha + 1}} $$

for \(x > 0\), \(\alpha > 0\) and \(\theta > 0\).

Cumulative distribution function (c.d.f):

$$ F(x) = 1 - \left(\frac{\theta}{x + \theta}\right)^\alpha $$

for \(x > 0\).

Moment generating function (m.g.f):

Does not exist in closed form, but the \(k\)th raw moment \(E[X^k]\) exists for \(-1 < k < \alpha\).

Note

There are many different definitions of the Pareto distribution in the literature; see Arnold (2015) or Kleiber and Kotz (2003). This implementation uses the Pareto distribution without a location parameter as described in actuar::Pareto.

References

Kleiber, C. and Kotz, S. (2003), Statistical Size Distributions in Economics and Actuarial Sciences, Wiley.

Klugman, S. A., Panjer, H. H. and Willmot, G. E. (2012), Loss Models, From Data to Decisions, Fourth Edition, Wiley.

See also

Examples

dist <- dist_pareto(shape = c(10, 3, 2, 1), scale = rep(1, 4))
dist
#> <distribution[4]>
#> [1] Pareto(10, 1) Pareto(3, 1)  Pareto(2, 1)  Pareto(1, 1) 

mean(dist)
#> [1] 0.1111111 0.5000000 1.0000000       Inf
variance(dist)
#> [1] 0.0154321 0.7500000       Inf       NaN
support(dist)
#> <support_region[4]>
#> [1] [0,Inf) [0,Inf) [0,Inf) [0,Inf)
generate(dist, 10)
#> [[1]]
#>  [1] 0.199464862 0.031031212 0.308510828 0.024651899 0.023364192 0.284767907
#>  [7] 0.017036675 0.082886923 0.015911397 0.005019324
#> 
#> [[2]]
#>  [1] 0.28478584 0.53188276 0.06426553 0.09304199 0.41834164 0.89052386
#>  [7] 0.43549444 0.01693421 0.46341002 0.94239762
#> 
#> [[3]]
#>  [1] 3.6882407 0.8170206 1.4097964 0.5915486 0.6534127 0.1652269 0.2499987
#>  [8] 0.5233614 0.3877956 0.6965226
#> 
#> [[4]]
#>  [1] 0.14837171 1.71118770 2.34585484 0.65074012 0.46148555 0.42638059
#>  [7] 0.08479179 0.32628692 1.13637539 0.25414272
#> 

density(dist, 2)
#> [1] 5.645029e-05 3.703704e-02 7.407407e-02 1.111111e-01
density(dist, 2, log = TRUE)
#> [1] -9.782150 -3.295837 -2.602690 -2.197225

cdf(dist, 4)
#> [1] 0.9999999 0.9920000 0.9600000 0.8000000

quantile(dist, 0.7)
#> [1] 0.1279449 0.4938016 0.8257419 2.3333333