Sampling distribution

The sampling distribution represents an empirical distribution based on observed samples. It is useful for bootstrapping, representing posterior distributions from Markov Chain Monte Carlo (MCMC) algorithms, or working with any empirical data where the parametric form is unknown. Unlike parametric distributions, the sampling distribution makes no assumptions about the underlying data-generating process and instead uses the sample itself to estimate distributional properties. The distribution can handle both univariate and multivariate samples.

dist_sample(x)

Arguments

x

A list of sampled values. For univariate distributions, each element should be a numeric vector. For multivariate distributions, each element should be a matrix where columns represent variables and rows represent observations.

Details

We recommend reading this documentation on pkgdown which renders math nicely. https://pkg.mitchelloharawild.com/distributional/reference/dist_sample.html

In the following, let \(X\) be a random variable with sample \(x_1, x_2, \ldots, x_n\) of size \(n\).

Support: The observed range of the sample

Mean (univariate):

$$ \bar{x} = \frac{1}{n} \sum_{i=1}^{n} x_i $$

Mean (multivariate): Computed independently for each variable.

Variance (univariate):

$$ s^2 = \frac{1}{n-1} \sum_{i=1}^{n} (x_i - \bar{x})^2 $$

Covariance (multivariate): The sample covariance matrix.

Skewness (univariate):

$$ g_1 = \frac{\sqrt{n} \sum_{i=1}^{n} (x_i - \bar{x})^3}{\left(\sum_{i=1}^{n} (x_i - \bar{x})^2\right)^{3/2}} \left(1 - \frac{1}{n}\right)^{3/2} $$

Probability density function: Approximated numerically using kernel density estimation.

Cumulative distribution function (univariate):

$$ F(q) = \frac{1}{n} \sum_{i=1}^{n} I(x_i \leq q) $$

where \(I(\cdot)\) is the indicator function.

Cumulative distribution function (multivariate):

$$ F(\mathbf{q}) = \frac{1}{n} \sum_{i=1}^{n} I(\mathbf{x}_i \leq \mathbf{q}) $$

where the inequality is applied element-wise.

Quantile function (univariate): The sample quantile, computed using the specified quantile type (see stats::quantile()).

Quantile function (multivariate): Marginal quantiles are computed independently for each variable.

Random generation: Bootstrap sampling with replacement from the empirical sample.

See also

Examples

# Univariate numeric samples
dist <- dist_sample(x = list(rnorm(100), rnorm(100, 10)))

dist
#> <distribution[2]>
#> [1] sample[100] sample[100]
mean(dist)
#> [1]  0.1244638 10.0087463
variance(dist)
#> [1] 1.2815536 0.9889088
skewness(dist)
#> [1] -0.3294489  0.3658181
generate(dist, 10)
#> [[1]]
#>  [1]  0.89681435  0.66172923  0.77796979 -0.57919735  0.44300601  0.11313826
#>  [7]  0.06372166 -0.79665963  0.93422264  0.05431970
#> 
#> [[2]]
#>  [1] 10.549839  9.583764  8.842354 10.592675 10.991759  9.865518 10.033976
#>  [8]  9.479471  9.043872 10.059874
#> 

density(dist, 1)
#> [1] 0.2692047 0.0000000

# Multivariate numeric samples
dist <- dist_sample(x = list(cbind(rnorm(100), rnorm(100, 10))))
dimnames(dist) <- c("x", "y")

dist
#> <distribution[1]>
#> [1] sample[100]
mean(dist)
#>              x        y
#> [1,] 0.1297508 10.03704
variance(dist)
#>                x           y
#> [1,]  0.99518690 -0.06661114
#> [2,] -0.06661114  1.00038342
generate(dist, 10)
#> [[1]]
#>                x         y
#>  [1,] -0.4409363 10.173137
#>  [2,] -0.3442147  9.753972
#>  [3,] -0.5753095  9.883991
#>  [4,] -0.1252417  8.353882
#>  [5,]  0.1803975  9.953664
#>  [6,]  0.5996725 10.296227
#>  [7,]  2.5753966  8.939885
#>  [8,]  0.5996725 10.296227
#>  [9,]  1.2076322  9.025916
#> [10,] -1.7987431 10.787037
#> 
quantile(dist, 0.4) # Returns the marginal quantiles
#>               x        y
#> [1,] -0.1623372 9.816357
cdf(dist, matrix(c(0.3,9), nrow = 1))
#> [1] 0.37