The Weibull distribution

dist_weibull(shape, scale)

Arguments

shape

shape and scale parameters, the latter defaulting to 1.

scale

shape and scale parameters, the latter defaulting to 1.

Details

Generalization of the gamma distribution. Often used in survival and time-to-event analyses.

We recommend reading this documentation on https://pkg.mitchelloharawild.com/distributional/, where the math will render nicely.

In the following, let \(X\) be a Weibull random variable with success probability p = \(p\).

Support: \(R^+\) and zero.

Mean: \(\lambda \Gamma(1+1/k)\), where \(\Gamma\) is the gamma function.

Variance: \(\lambda [ \Gamma (1 + \frac{2}{k} ) - (\Gamma(1+ \frac{1}{k}))^2 ]\)

Probability density function (p.d.f):

$$ f(x) = \frac{k}{\lambda}(\frac{x}{\lambda})^{k-1}e^{-(x/\lambda)^k}, x \ge 0 $$

Cumulative distribution function (c.d.f):

$$F(x) = 1 - e^{-(x/\lambda)^k}, x \ge 0$$

Moment generating function (m.g.f):

$$\sum_{n=0}^\infty \frac{t^n\lambda^n}{n!} \Gamma(1+n/k), k \ge 1$$

See also

Examples

dist <- dist_weibull(shape = c(0.5, 1, 1.5, 5), scale = rep(1, 4))

dist
#> <distribution[4]>
#> [1] Weibull(0.5, 1) Weibull(1, 1)   Weibull(1.5, 1) Weibull(5, 1)  
mean(dist)
#> [1] 2.0000000 1.0000000 0.9027453 0.9181687
variance(dist)
#> [1] 20.00000000  1.00000000  0.37569028  0.04422998
skewness(dist)
#> [1]   5.0737409   0.0000000  -0.7764597 -38.1055455
kurtosis(dist)
#> [1]  87.48382  14.00000  12.28013 660.88795

generate(dist, 10)
#> [[1]]
#>  [1] 0.87458151 0.26631695 0.09613336 7.20846311 0.47140740 0.11897225
#>  [7] 0.01731281 2.26590036 0.20200736 0.07669760
#> 
#> [[2]]
#>  [1] 0.9615574 0.6990878 0.5067366 0.8706683 1.7161692 0.5478480 0.2296818
#>  [8] 2.3261885 0.3555478 2.1239887
#> 
#> [[3]]
#>  [1] 0.7274676 1.4573269 1.2804433 0.7761788 0.7286180 2.3020933 0.8050590
#>  [8] 0.4377178 1.3489056 0.5391107
#> 
#> [[4]]
#>  [1] 0.8342974 0.8657350 1.0873400 0.9511967 1.0350576 0.6277494 0.8763944
#>  [8] 1.0489659 1.1447902 0.8620448
#> 

density(dist, 2)
#> [1] 8.595475e-02 1.353353e-01 1.253822e-01 1.013133e-12
density(dist, 2, log = TRUE)
#> [1]  -2.453934  -2.000000  -2.076388 -27.617973

cdf(dist, 4)
#> [1] 0.8646647 0.9816844 0.9996645 1.0000000

quantile(dist, 0.7)
#> [1] 1.449551 1.203973 1.131734 1.037823