The Weibull distribution

Generalization of the gamma distribution. Often used in survival and time-to-event analyses.

dist_weibull(shape, scale)

Arguments

shape, scale

shape and scale parameters, the latter defaulting to 1.

Details

We recommend reading this documentation on https://pkg.mitchelloharawild.com/distributional/, where the math will render nicely.

In the following, let \(X\) be a Weibull random variable with success probability p = \(p\).

Support: \(R^+\) and zero.

Mean: \(\lambda \Gamma(1+1/k)\), where \(\Gamma\) is the gamma function.

Variance: \(\lambda [ \Gamma (1 + \frac{2}{k} ) - (\Gamma(1+ \frac{1}{k}))^2 ]\)

Probability density function (p.d.f):

$$ f(x) = \frac{k}{\lambda}(\frac{x}{\lambda})^{k-1}e^{-(x/\lambda)^k}, x \ge 0 $$

Cumulative distribution function (c.d.f):

$$F(x) = 1 - e^{-(x/\lambda)^k}, x \ge 0$$

Moment generating function (m.g.f):

$$\sum_{n=0}^\infty \frac{t^n\lambda^n}{n!} \Gamma(1+n/k), k \ge 1$$

See also

Examples

dist <- dist_weibull(shape = c(0.5, 1, 1.5, 5), scale = rep(1, 4))

dist
#> <distribution[4]>
#> [1] Weibull(0.5, 1) Weibull(1, 1)   Weibull(1.5, 1) Weibull(5, 1)  
mean(dist)
#> [1] 2.0000000 1.0000000 0.9027453 0.9181687
variance(dist)
#> [1] 20.00000000  1.00000000  0.37569028  0.04422998
skewness(dist)
#> [1]  6.6187612  2.0000000  1.0719866 -0.2541096
kurtosis(dist)
#> [1] 84.7200000  6.0000000  1.3904036 -0.1197099

generate(dist, 10)
#> [[1]]
#>  [1] 0.60561454 2.31744364 1.82899393 0.04055978 1.07558476 0.05996703
#>  [7] 0.39543432 0.84619154 0.03914828 4.66465783
#> 
#> [[2]]
#>  [1] 0.9217163 1.8950802 0.2866091 1.2502927 1.0870614 1.8048753 0.1595849
#>  [8] 3.0363772 2.7971648 0.1010004
#> 
#> [[3]]
#>  [1] 1.5405083 1.6740989 2.7108842 0.3421950 0.9858303 0.8584267 1.9198059
#>  [8] 0.7277771 0.8448059 0.2102344
#> 
#> [[4]]
#>  [1] 1.0814816 0.6704181 0.8852123 0.8712513 0.8197815 0.7908478 1.1324983
#>  [8] 1.2325753 0.6696045 0.6393683
#> 

density(dist, 2)
#> [1] 8.595475e-02 1.353353e-01 1.253822e-01 1.013133e-12
density(dist, 2, log = TRUE)
#> [1]  -2.453934  -2.000000  -2.076388 -27.617973

cdf(dist, 4)
#> [1] 0.8646647 0.9816844 0.9996645 1.0000000

quantile(dist, 0.7)
#> [1] 1.449551 1.203973 1.131734 1.037823