Create a distribution from p/d/q/r style functions

dist_wrap(dist, ..., package = "stats")

Arguments

dist
The name of the distribution used in the functions (name that is prefixed by p/d/q/r)

...
Named arguments used to parameterise the distribution.

package
The package from which the distribution is provided.

Details

If a distribution is not yet supported, you can vectorise p/d/q/r functions using this function. dist_wrap() stores the distributions parameters, and provides wrappers which call the appropriate p/d/q/r functions.

Using this function to wrap a distribution should only be done if the distribution is not yet available in this package. If you need a distribution which isn't in the package yet, consider making a request at https://github.com/mitchelloharawild/distributional/issues.

Examples

dist <- dist_wrap("norm", mean = 1:3, sd = c(3, 9, 2))

density(dist, 1) # dnorm()

#> [1] 0.13298076 0.04405414 0.12098536

cdf(dist, 4) # pnorm()

#> [1] 0.8413447 0.5879296 0.6914625

quantile(dist, 0.975) # qnorm()

#> [1]  6.879892 19.639676  6.919928

generate(dist, 10) # rnorm()

#> [[1]]
#>  [1]  2.2198996  1.5486744 -4.6301461 -0.5322246  0.5259997  4.6454550
#>  [7]  1.1266567  2.1922211  4.8715095  3.6000141
#> 
#> [[2]]
#>  [1]   7.5453880 -11.3430179   3.2256863  13.6548980  -9.6158501  -6.7938406
#>  [7]  16.2245056   0.2115309   3.6040012  10.4841796
#> 
#> [[3]]
#>  [1]  1.5534312  2.2070353  3.6316605  0.3906120 -0.6187544  1.1633196
#>  [7]  3.6561224  3.3957038  0.7122406  2.8709851
#> 

library(actuar)

#> 
#> Attaching package: ‘actuar’

#> The following object is masked from ‘package:grDevices’:
#> 
#>     cm

dist <- dist_wrap("invparalogis", package = "actuar", shape = 2, rate = 2)
density(dist, 1) # actuar::dinvparalogis()

#> [1] 0.512

cdf(dist, 4) # actuar::pinvparalogis()

#> [1] 0.9694675

quantile(dist, 0.975) # actuar::qinvparalogis()

#> [1] 4.429922

generate(dist, 10) # actuar::rinvparalogis()

#> [[1]]
#>  [1] 0.4883214 0.5091362 0.2183942 0.4245982 0.9619228 0.4708693 0.2931011
#>  [8] 1.0096480 0.5630054 0.9977921
#>