Create a distribution from p/d/q/r style functions

dist_wrap(dist, ..., package = NULL)

Arguments

dist

The name of the distribution used in the functions (name that is prefixed by p/d/q/r)

...

Named arguments used to parameterise the distribution.

package

The package from which the distribution is provided. If NULL, the calling environment's search path is used to find the distribution functions. Alternatively, an arbitrary environment can also be provided here.

Details

If a distribution is not yet supported, you can vectorise p/d/q/r functions using this function. dist_wrap() stores the distributions parameters, and provides wrappers which call the appropriate p/d/q/r functions.

Using this function to wrap a distribution should only be done if the distribution is not yet available in this package. If you need a distribution which isn't in the package yet, consider making a request at https://github.com/mitchelloharawild/distributional/issues.

Examples

dist <- dist_wrap("norm", mean = 1:3, sd = c(3, 9, 2))

density(dist, 1) # dnorm()
#> [1] 0.13298076 0.04405414 0.12098536
cdf(dist, 4) # pnorm()
#> [1] 0.8413447 0.5879296 0.6914625
quantile(dist, 0.975) # qnorm()
#> [1]  6.879892 19.639676  6.919928
generate(dist, 10) # rnorm()
#> [[1]]
#>  [1] -0.8367252  3.0377620  2.4604545  3.1374014  1.9177305 -0.6550399
#>  [7]  0.1300628 -0.4067899  1.6288172 -0.8525970
#> 
#> [[2]]
#>  [1]   6.9176576   2.9228457   2.5577639 -13.9758194   8.3430013   8.1398837
#>  [7]   0.7695594  -4.0600188   4.9625839   8.3129567
#> 
#> [[3]]
#>  [1]  1.8799803  5.2585107 -0.3316427  5.4183372  3.4500007  4.2314959
#>  [7]  3.9696255  1.4751214  3.9903347  0.3461239
#> 

library(actuar)
#> 
#> Attaching package: ‘actuar’
#> The following objects are masked from ‘package:stats’:
#> 
#>     sd, var
#> The following object is masked from ‘package:grDevices’:
#> 
#>     cm
dist <- dist_wrap("invparalogis", package = "actuar", shape = 2, rate = 2)
density(dist, 1) # actuar::dinvparalogis()
#> [1] 0.512
cdf(dist, 4) # actuar::pinvparalogis()
#> [1] 0.9694675
quantile(dist, 0.975) # actuar::qinvparalogis()
#> [1] 4.429922
generate(dist, 10) # actuar::rinvparalogis()
#> [[1]]
#>  [1] 3.1024189 0.5442053 1.2128388 2.8898399 0.1278556 0.6653644 0.1272430
#>  [8] 0.8390043 1.1889970 0.8438120
#>